refactor: use DetailPageLayout for all dao pages
