+++ apps/main/src/dao/components/PageAnalytics/HoldersTable/index.tsx
@@ -6,7 +6,7 @@ import { useStore } from '@/dao/store/useStore'
 import type { AllHoldersSortBy } from '@/dao/types/dao.types'
 import { getEthPath } from '@/dao/utils'
 import type { Locker } from '@curvefi/prices-api/dao'
-import { Box } from '@ui/Box'
+import Stack from '@mui/material/Stack'
 import { formatNumber, formatDate } from '@ui/utils'
 import { t } from '@ui-kit/lib/i18n'
 import { DAO_ROUTES } from '@ui-kit/shared/routes'
@@ -31,7 +31,7 @@ export const TopHoldersTable = () => {
   ]
 
   return (
-    <Wrapper>
+    <Stack sx={{ backgroundColor: t => t.design.Layer[1].Fill }}>
       <PaginatedTable
         data={holdersArray}
         minWidth={tableMinWidth}
@@ -73,16 +73,9 @@ export const TopHoldersTable = () => {
           </TableRowWrapper>
         )}
       />
-    </Wrapper>
+    </Stack>
   )
 }
-
-const Wrapper = styled(Box)`
-  display: flex;
-  flex-direction: column;
-  padding-bottom: var(--spacing-3);
-  width: 100%;
-`
 
 const StyledTableDataLink = styled(TableDataLink)`
   text-decoration: none;
