+++ apps/main/src/dao/components/PageGauge/index.tsx
@@ -1,20 +1,20 @@
-import { styled } from 'styled-components'
 import { GaugeWeightHistoryChart } from '@/dao/components/Charts/GaugeWeightHistoryChart'
 import { useGauges } from '@/dao/queries/gauges.query'
 import type { GaugeUrlParams } from '@/dao/types/dao.types'
 import { getEthPath } from '@/dao/utils'
-import { Box } from '@ui/Box'
+import Stack from '@mui/material/Stack'
+import { useParams } from '@ui-kit/hooks/router'
+import { t } from '@ui-kit/lib/i18n'
 import { DAO_ROUTES } from '@ui-kit/shared/routes'
+import { PAGE_SPACING } from '@ui-kit/widgets/DetailPageLayout/constants'
+import { DetailPageLayout } from '@ui-kit/widgets/DetailPageLayout/DetailPageLayout'
 import { BackButton } from '../BackButton'
 import { GaugeHeader } from './GaugeHeader'
 import { GaugeMetrics } from './GaugeMetrics'
 import { GaugeVotesTable } from './GaugeVotesTable'
 
-type GaugeProps = {
-  routerParams: GaugeUrlParams
-}
-
-export const Gauge = ({ routerParams: { gaugeAddress: rGaugeAddress } }: GaugeProps) => {
+export const Gauge = () => {
+  const { gaugeAddress: rGaugeAddress } = useParams<GaugeUrlParams>()
   const gaugeAddress = rGaugeAddress.toLowerCase()
   const { data: gauges, isLoading: gaugesIsLoading } = useGauges({})
   const gaugeData = gauges?.[gaugeAddress]
@@ -22,41 +22,14 @@ export const Gauge = ({ routerParams: { gaugeAddress: rGaugeAddress } }: GaugePr
   const tableMinWidth = 21.875
 
   return (
-    <Wrapper>
-      <BackButton path={getEthPath(DAO_ROUTES.PAGE_GAUGES)} label="Back to gauges" />
-      <GaugePageContainer variant="secondary">
+    <DetailPageLayout formTabs={null}>
+      <BackButton path={getEthPath(DAO_ROUTES.PAGE_GAUGES)} label={t`Back to gauges`} />
+      <Stack gap={PAGE_SPACING} sx={{ backgroundColor: t => t.design.Layer[1].Fill }}>
         <GaugeHeader gaugeData={gaugeData} dataLoading={gaugesIsLoading} />
         <GaugeMetrics gaugeData={gaugeData} dataLoading={gaugesIsLoading} />
-        <Content>{gaugeData && <GaugeWeightHistoryChart gaugeAddress={gaugeData.address} minHeight={25} />}</Content>
+        <Stack>{gaugeData && <GaugeWeightHistoryChart gaugeAddress={gaugeData.address} minHeight={25} />}</Stack>
         {gaugeData && <GaugeVotesTable gaugeAddress={gaugeData.address} tableMinWidth={tableMinWidth} />}
-      </GaugePageContainer>
-    </Wrapper>
+      </Stack>
+    </DetailPageLayout>
   )
 }
-
-const Wrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  margin: var(--spacing-4) auto var(--spacing-6);
-  width: 65rem;
-  max-width: 100%;
-  flex-grow: 1;
-  min-height: 100%;
-  @media (min-width: 34.375rem) {
-    max-width: 95%;
-  }
-`
-
-const GaugePageContainer = styled(Box)`
-  display: flex;
-  flex-direction: column;
-  margin-bottom: auto;
-  width: 100%;
-`
-
-const Content = styled.div`
-  display: flex;
-  flex-direction: column;
-  padding: var(--spacing-3);
-  width: 100%;
-`
