+++ apps/main/src/dao/components/PageUser/UserGaugeVotesTable/index.tsx
@@ -10,6 +10,7 @@ import {
 import { useStore } from '@/dao/store/useStore'
 import { SortDirection, UserGaugeVotesSortBy } from '@/dao/types/dao.types'
 import { getEthPath } from '@/dao/utils'
+import Box from '@mui/material/Box'
 import { formatLocaleDateFromTimestamp } from '@ui/utils'
 import { t } from '@ui-kit/lib/i18n'
 import { DAO_ROUTES } from '@ui-kit/shared/routes'
@@ -42,35 +43,37 @@ export const UserGaugeVotesTable = ({ userAddress, tableMinWidth }: UserGaugeVot
   )
 
   return (
-    <PaginatedTable<UserGaugeVoteFormatted>
-      data={sortedUserGaugeVotes}
-      minWidth={tableMinWidth}
-      isLoading={isLoading}
-      isError={isError}
-      isSuccess={isSuccess}
-      columns={GAUGE_VOTES_LABELS}
-      sortBy={userGaugeVotesSortBy}
-      errorMessage={t`An error occurred while fetching user gauge votes.`}
-      setSortBy={key => setUserGaugeVotesSortBy(key as UserGaugeVotesSortBy)}
-      getData={() => invalidateUserGaugeVoteQuery({ userAddress })}
-      noDataMessage={t`No gauge votes found for this user.`}
-      gridTemplateColumns={gridTemplateColumns}
-      renderRow={(gaugeVote, index) => (
-        <TableRowWrapper key={index} columns={GAUGE_VOTES_LABELS.length} gridTemplateColumns={gridTemplateColumns}>
-          <TableData className="align-left">{gaugeVote.gaugeName}</TableData>
-          <TableData className={`right-padding ${userGaugeVotesSortBy.key === 'timestamp' ? 'sortby-active' : ''}`}>
-            {formatLocaleDateFromTimestamp(gaugeVote.timestamp)}
-          </TableData>
-          <TableData
-            className={userGaugeVotesSortBy.key === 'weight' ? 'sortby-active right-padding' : 'right-padding'}
-          >
-            {formatNumber(gaugeVote.weight / 100, { unit: 'percentage', abbreviate: false })}
-          </TableData>
-          <TableDataLink href={getEthPath(`${DAO_ROUTES.PAGE_GAUGES}/${gaugeVote.gauge}`)} className="right-padding">
-            {shortenAddress(gaugeVote.gauge)}
-          </TableDataLink>
-        </TableRowWrapper>
-      )}
-    />
+    <Box sx={{ backgroundColor: t => t.design.Layer[1].Fill }}>
+      <PaginatedTable<UserGaugeVoteFormatted>
+        data={sortedUserGaugeVotes}
+        minWidth={tableMinWidth}
+        isLoading={isLoading}
+        isError={isError}
+        isSuccess={isSuccess}
+        columns={GAUGE_VOTES_LABELS}
+        sortBy={userGaugeVotesSortBy}
+        errorMessage={t`An error occurred while fetching user gauge votes.`}
+        setSortBy={key => setUserGaugeVotesSortBy(key as UserGaugeVotesSortBy)}
+        getData={() => invalidateUserGaugeVoteQuery({ userAddress })}
+        noDataMessage={t`No gauge votes found for this user.`}
+        gridTemplateColumns={gridTemplateColumns}
+        renderRow={(gaugeVote, index) => (
+          <TableRowWrapper key={index} columns={GAUGE_VOTES_LABELS.length} gridTemplateColumns={gridTemplateColumns}>
+            <TableData className="align-left">{gaugeVote.gaugeName}</TableData>
+            <TableData className={`right-padding ${userGaugeVotesSortBy.key === 'timestamp' ? 'sortby-active' : ''}`}>
+              {formatLocaleDateFromTimestamp(gaugeVote.timestamp)}
+            </TableData>
+            <TableData
+              className={userGaugeVotesSortBy.key === 'weight' ? 'sortby-active right-padding' : 'right-padding'}
+            >
+              {formatNumber(gaugeVote.weight / 100, { unit: 'percentage', abbreviate: false })}
+            </TableData>
+            <TableDataLink href={getEthPath(`${DAO_ROUTES.PAGE_GAUGES}/${gaugeVote.gauge}`)} className="right-padding">
+              {shortenAddress(gaugeVote.gauge)}
+            </TableDataLink>
+          </TableRowWrapper>
+        )}
+      />
+    </Box>
   )
 }
