+++ packages/curve-ui-kit/src/widgets/DetailPageLayout/DetailPageLayout.tsx
@@ -55,11 +55,13 @@ export const DetailPageLayout = ({
   header,
   children,
   footer,
+  testId,
 }: {
   formTabs: ReactNode
   header?: ReactNode
   children?: ReactNode
   footer?: ReactNode
+  testId?: string
 }) => {
   const navHeight = useLayoutStore(state => state.navHeight)
   const isNewLayout = useRightFormTabsLayout()
@@ -77,7 +79,7 @@ export const DetailPageLayout = ({
   return (
     <Grid
       container
-      data-testid="detail-page-layout"
+      data-testid={testId ?? 'detail-page-layout'}
       spacing={PAGE_SPACING}
       sx={PAGE_MARGIN}
       direction={isNewLayout ? 'row-reverse' : 'row'}
